+++ packages/tinacms/package.json
@@ -9,7 +9,8 @@
     "entryPoints": [
       "src/index.ts",
       "src/edit-state.tsx",
-      "src/rich-text.tsx"
+      "src/rich-text.tsx",
+      "src/client.ts"
     ]
   },
   "typings": "dist/index.d.ts",
@@ -28,6 +29,7 @@
     "@tinacms/sharedctx": "workspace:*",
     "@tinacms/toolkit": "workspace:*",
     "crypto-js": "^4.0.0",
+    "fetch-ponyfill": "^7.1.0",
     "final-form": "4.20.1",
     "graphql": "^15.1.0",
     "graphql-tag": "^2.11.0",
