+++ packages/tinacms/src/index.ts
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-export * from './client'
+export * from './internalClient'
 export * from './auth'
 export * from './utils'
 export * from './tina-cms'
